+++ src/types/supabase.ts
@@ -32,18 +32,28 @@ export interface Token {
   max_supply: number;
 }
 
+export type SortField =
+  | "price"
+  | "price24hChanges"
+  | "marketCap"
+  | "created_at";
+
 export interface DAO {
   id: string;
-  created_at: string;
   name: string;
   mission: string;
   description: string;
   image_url: string;
   is_graduated: boolean;
   is_deployed: boolean;
-  x_url?: string;
-  telegram_url?: string;
-  website_url?: string;
+  created_at: string;
+  author_id: string;
+  user_id?: string;
+  extensions?: Array<{
+    id: string;
+    type: string;
+    contract_principal?: string;
+  }>;
 }
 
 export interface Holder {
